fix color provider rendering twice
