+++ src/components/color/store.ts
@@ -1,8 +1,7 @@
 import { createStore } from "zustand";
-import { getRandomColor } from "~/lib/color";
 import type { ColorSpace, ColorAction } from "~/lib/types";
 
-export type ColorState = ColorSpace & ColorAction;
+type ColorState = ColorSpace & ColorAction;
 
 export function createColorStore(initValue: ColorSpace) {
   return createStore<ColorState>()((set) => ({
@@ -14,15 +13,3 @@ export function createColorStore(initValue: ColorSpace) {
       })),
   }));
 }
-
-export const initColorStore = () => {
-  const color = getRandomColor();
-
-  return {
-    raw: color.toRgb(),
-    hex: color.toHex(),
-    hsl: color.toHsl(),
-    hsv: color.toHsv(),
-    rgb: color.toRgb(),
-  };
-};
